+++ src/main/java/org/study/entities/databases/OracleDatabase.java
@@ -7,13 +7,13 @@ public class OracleDatabase<T> implements Database<T> {
     }
 
     @Override
-    public T remove(int id) {
+    public T remove(Long id) {
         System.out.println("Remove obj id: " + id);
         return null;
     }
 
     @Override
-    public T select(int id) {
+    public T select(Long id) {
         System.out.println("Select obj id: " + id);
         return null;
     }
